feat(build): vastly simplify the build system
