+++ assets/completions/_zellij
@@ -20,6 +20,7 @@ _zellij() {
 '-o+[Send "open file in new pane" to active zellij session]' \
 '--open-file=[Send "open file in new pane" to active zellij session]' \
 '--max-panes=[Maximum panes on screen, caution: opening more panes will close old ones]' \
+'--data-dir=[Change where zellij looks for layouts and plugins]' \
 '-l+[Path to a layout yaml file]' \
 '--layout=[Path to a layout yaml file]' \
 '-m[Send "move focused pane" to active zellij session]' \
