+++ src/cli.rs
@@ -20,6 +20,10 @@ pub struct CliArgs {
     #[structopt(long)]
     pub max_panes: Option<usize>,
 
+    /// Change where zellij looks for layouts and plugins
+    #[structopt(long)]
+    pub data_dir: Option<PathBuf>,
+
     /// Path to a layout yaml file
     #[structopt(short, long)]
     pub layout: Option<PathBuf>,
